+++ admin/users.php
@@ -1,5 +1,5 @@
 <?php
-// Include header
+// Include the header
 require_once __DIR__.'/header.php';
 
 // Create a User object
@@ -16,15 +16,15 @@ $rs_user = new User;
 	switch($action) {
 		case 'create':
 			// Create a new user
-			userHasPrivilege($_SESSION['role'], 'can_create_users') ? $rs_user->createEntry() : redirect('users.php');
+			userHasPrivilege($_SESSION['role'], 'can_create_users') ? $rs_user->createUser() : redirect('users.php');
 			break;
 		case 'edit':
 			// Edit an existing user
-			userHasPrivilege($_SESSION['role'], 'can_edit_users') ? $rs_user->editEntry($id) : redirect('users.php');
+			userHasPrivilege($_SESSION['role'], 'can_edit_users') ? $rs_user->editUser($id) : redirect('users.php');
 			break;
 		case 'delete':
 			// Delete an existing user
-			userHasPrivilege($_SESSION['role'], 'can_delete_users') ? $rs_user->deleteEntry($id) : redirect('users.php');
+			userHasPrivilege($_SESSION['role'], 'can_delete_users') ? $rs_user->deleteUser($id) : redirect('users.php');
 			break;
 		case 'reset_password':
 			// Reset a user's password
@@ -32,10 +32,10 @@ $rs_user = new User;
 			break;
 		default:
 			// List all users
-			userHasPrivilege($_SESSION['role'], 'can_view_users') ? $rs_user->listEntries() : redirect('index.php');
+			userHasPrivilege($_SESSION['role'], 'can_view_users') ? $rs_user->listUsers() : redirect('index.php');
 	}
 	?>
 </div>
 <?php
-// Include footer
+// Include the footer
 require_once __DIR__.'/footer.php';