+++ admin/widgets.php
@@ -1,5 +1,5 @@
 <?php
-// Include header
+// Include the header
 require_once __DIR__.'/header.php';
 
 // Create a Widget object
@@ -16,22 +16,22 @@ $rs_widget = new Widget;
 	switch($action) {
 		case 'create':
 			// Create a new widget
-			userHasPrivilege($_SESSION['role'], 'can_create_widgets') ? $rs_widget->createEntry() : redirect('widgets.php');
+			userHasPrivilege($_SESSION['role'], 'can_create_widgets') ? $rs_widget->createWidget() : redirect('widgets.php');
 			break;
 		case 'edit':
 			// Edit an existing widget
-			userHasPrivilege($_SESSION['role'], 'can_edit_widgets') ? $rs_widget->editEntry($id) : redirect('widgets.php');
+			userHasPrivilege($_SESSION['role'], 'can_edit_widgets') ? $rs_widget->editWidget($id) : redirect('widgets.php');
 			break;
 		case 'delete':
 			// Delete an existing widget
-			userHasPrivilege($_SESSION['role'], 'can_delete_widgets') ? $rs_widget->deleteEntry($id) : redirect('widgets.php');
+			userHasPrivilege($_SESSION['role'], 'can_delete_widgets') ? $rs_widget->deleteWidget($id) : redirect('widgets.php');
 			break;
 		default:
 			// List all widgets
-			userHasPrivilege($_SESSION['role'], 'can_view_widgets') ? $rs_widget->listEntries() : redirect('index.php');
+			userHasPrivilege($_SESSION['role'], 'can_view_widgets') ? $rs_widget->listWidgets() : redirect('index.php');
 	}
 	?>
 </div>
 <?php
-// Include footer
+// Include the footer
 require_once __DIR__.'/footer.php';